requirements.txt
flask>=2.0.1
ultralytics>=8.0.0
opencv-python>=4.5.3
numpy>=1.21.0
PyYAML>=5.4.1
pillow>=8.3.1
torch>=1.9.0
torchvision>=0.10.0